@peterstace 2024-04-18
> @peterstace not sure if a bug fix needs to go into the release notes ?

I'll add something to the changelog after this goes into master.